Get array items with array_walk, and returns indices for each element.
Bash 5.2 introduced more changes in how arrays are represented internally. Such
representation can be modified at compile-time[1], so it is not a valid approach
anymore to assume a specific representation.

This commit changes the way array items are read. Now, it uses the `array_walk`
function from Bash to iterate over array elements, so it should be compatible
with all recent bash versions.

The elements of the array are now returned as a tuple of `(index, value)`, since
arrays in Bash can contain holes. This is a breaking change.

[1]: With the `--enable-alt-array-implementation` configure option.