Generate `long_doc` field as a single string.
In #4 it was discovered that empty strings in `long_doc` can produce unexpected
messages if Bash is running with a non-English locale.

To avoid the problem, this commit changes the generation of the `long_doc` field
to be an array of a single element, instead of an element per line. This is the
same method used for builtins included in Bash.

The `--help` output is identical to the previous method.